delete pkt_array and move to vector
